Rename of utils.js as innerLogicFunctions.js
