+++ src/innerLogicFunctions.js
